+++ include/oneapi/dpl/pstl/hetero/dpcpp/unseq_backend_sycl.h
@@ -711,41 +711,42 @@ struct __copy_by_mask
     {
         using std::get;
         auto __item_idx = __item.get_linear_id();
-        if (__item_idx < __n && get<N>(__in_acc[__item_idx]))
+
+        auto __local_scan_of_mask = (__item_idx < __n) ? get<N>(__in_acc[__item_idx]) : 0;
+        // Restore the mask from the scan: for the first element in a group it is equal to the scan value,
+        // then to the difference with the previous value.
+        bool __restored_mask = __local_scan_of_mask > 0 &&
+            (__item_idx % __size_per_wg == 0 || __local_scan_of_mask > get<N>(__in_acc[__item_idx - 1]));
+        if (__restored_mask)
         {
-            auto __out_idx = get<N>(__in_acc[__item_idx]) - 1;
-
-            // If we work with tuples we might have a situation when internal tuple is assigned to std::tuple
-            // (e.g. returned by user-provided lambda).
-            // For internal::tuple<T...> we have a conversion operator to std::tuple<T...>. The problem here
-            // is that the types of these 2 tuples may be different but still convertible to each other.
-            // Technically this should be solved by adding to internal::tuple<T...> an additional conversion
-            // operator to std::tuple<U...>, but for some reason this doesn't work(conversion from
-            // std::tuple<T...> to std::tuple<U...> fails). What does work is the explicit cast:
-            // for internal::tuple<T...> we define a field that provides a corresponding std::tuple<T...>
-            // with matching types. We get this type(see __tuple_type definition below) and use it
-            // for static cast to explicitly convert internal::tuple<T...> -> std::tuple<T...>.
-            // Now we have the following assignment std::tuple<U...> = std::tuple<T...> which works as expected.
-            // NOTE: we only need this explicit conversion when we have internal::tuple and
-            // std::tuple as operands, in all the other cases this is not necessary and no conversion
-            // is performed(i.e. __tuple_type is the same type as its operand).
-            using __tuple_type =
-                typename __internal::__get_tuple_type<std::decay_t<decltype(get<0>(__in_acc[__item_idx]))>,
-                                                      std::decay_t<decltype(__out_acc[__out_idx])>>::__type;
-
-            // calculation of position for copy
+            // calculate the output position
+            auto __out_idx = __local_scan_of_mask - 1;
             if (__item_idx >= __size_per_wg)
             {
                 auto __wg_sums_idx = __item_idx / __size_per_wg - 1;
                 __out_idx += __wg_sums_ptr[__wg_sums_idx];
             }
-            if (__item_idx % __size_per_wg == 0 || (get<N>(__in_acc[__item_idx]) != get<N>(__in_acc[__item_idx - 1])))
-            {
-                if (__out_idx < __n_out)
-                    __assigner(static_cast<__tuple_type>(get<0>(__in_acc[__item_idx])), __out_acc[__out_idx]);
-                if (__out_idx == __n_out)
-                    __ret_ptr[1] = __item_idx;
-            }
+            // If we work with tuples we might have a situation when internal tuple is assigned to std::tuple
+            // (e.g. returned by user-provided lambda).
+            // For internal::tuple<T...> we have a conversion operator to std::tuple<T...>. The problem here
+            // is that the types of these 2 tuples may be different but still convertible to each other.
+            // Technically this should be solved by adding to internal::tuple<T...> an additional conversion
+            // operator to std::tuple<U...>, but for some reason this doesn't work (conversion from
+            // std::tuple<T...> to std::tuple<U...> fails). What does work is the explicit cast:
+            // for internal::tuple<T...> we define a field that provides a corresponding std::tuple<T...>
+            // with matching types. We get this type (see __tuple_type definition below) and use it
+            // for static cast to explicitly convert internal::tuple<T...> -> std::tuple<T...>.
+            // Now we have the following assignment std::tuple<U...> = std::tuple<T...> which works as expected.
+            // NOTE: we only need this explicit conversion when we have internal::tuple and
+            // std::tuple as operands, in all the other cases this is not necessary and no conversion
+            // is performed (i.e. __tuple_type is the same type as its operand).
+            using __tuple_type =
+                typename __internal::__get_tuple_type<std::decay_t<decltype(get<0>(__in_acc[__item_idx]))>,
+                                                      std::decay_t<decltype(__out_acc[__out_idx])>>::__type;
+            if (__out_idx < __n_out)
+                __assigner(static_cast<__tuple_type>(get<0>(__in_acc[__item_idx])), __out_acc[__out_idx]);
+            if (__out_idx == __n_out)
+                __ret_ptr[1] = __item_idx;
         }
     }
 };
@@ -858,23 +859,42 @@ struct __scan
     _GlobalAssigner __gl_assigner;
     _DataAccessor __data_acc;
 
-    template <typename _NDItemId, typename _Size, typename _AccLocal, typename _InAcc, typename _OutAcc,
-              typename _WGSumsPtr, typename _SizePerWG, typename _WGSize, typename _ItersPerWG>
-    void
-    scan_impl(_NDItemId __item, _Size __n, _Size __n_out, _AccLocal& __local_acc, const _InAcc& __acc,
-              _OutAcc& __out_acc, _WGSumsPtr* __wg_sums_ptr, _SizePerWG __size_per_wg, _WGSize __wgroup_size,
-              _ItersPerWG __iters_per_wg, _InitType __init, std::false_type /*has_known_identity*/) const
+    // A workaround implementation of inclusive scan-over-group
+    template <typename _AccLocal>
+    _Tp
+    __log_scan_over_group(std::size_t __wgroup_size, sycl::nd_item<1> __item, std::size_t __local_id,
+                          _AccLocal& __local_acc, _Tp __adder) const
     {
-        ::std::size_t __group_id = __item.get_group(0);
-        ::std::size_t __global_id = __item.get_global_id(0);
-        ::std::size_t __local_id = __item.get_local_id(0);
+        _Tp __value = __local_acc[__local_id];
+        __dpl_sycl::__group_barrier(__item);
+        // Hillis-Steele algorithm
+        for (std::size_t __shift = 1; __shift < __wgroup_size; __shift *= 2)
+        {
+            if (__local_id >= __shift)
+                __value = __bin_op(__local_acc[__local_id - __shift], __value);
+            __dpl_sycl::__group_barrier(__item);
+            __local_acc[__local_id] = __value;
+            __dpl_sycl::__group_barrier(__item);
+        }
+        return __bin_op(__adder, __value);
+    }
+
+    template <typename _Size, typename _AccLocal, typename _InAcc, typename _OutAcc, typename _WGSumsPtr>
+    void
+    __scan_impl(sycl::nd_item<1> __item, _Size __n, _Size __n_out, _AccLocal& __local_acc, const _InAcc& __acc,
+                _OutAcc& __out_acc, _WGSumsPtr* __wg_sums_ptr, std::size_t __size_per_wg, std::size_t __wgroup_size,
+                std::size_t __iters_per_wg, _InitType __init, std::false_type /*has_known_identity*/) const
+    {
+        std::size_t __group_id = __item.get_group(0);
+        std::size_t __global_id = __item.get_global_id(0);
+        std::size_t __local_id = __item.get_local_id(0);
         __init_processing<_Tp> __use_init{};
 
-        constexpr ::std::size_t __shift = _Inclusive{} ? 0 : 1;
+        constexpr std::size_t __shift = _Inclusive{} ? 0 : 1;
 
-        ::std::size_t __adjusted_global_id = __local_id + __size_per_wg * __group_id;
+        std::size_t __adjusted_global_id = __local_id + __size_per_wg * __group_id;
         auto __adder = __local_acc[0];
-        for (_ItersPerWG __iter = 0; __iter < __iters_per_wg; ++__iter, __adjusted_global_id += __wgroup_size)
+        for (std::size_t __iter = 0; __iter < __iters_per_wg; ++__iter, __adjusted_global_id += __wgroup_size)
         {
             if (__adjusted_global_id < __n)
             {
@@ -889,7 +909,7 @@ struct __scan
                 __use_init(__init, __local_acc[__global_id], __bin_op);
 
             // 1. reduce
-            ::std::size_t __k = 1;
+            std::size_t __k = 1;
             // TODO: use adjacent work items for better SIMD utilization
             // Consider the example with the mask of work items performing reduction:
             // iter    now         proposed
@@ -914,7 +934,7 @@ struct __scan
             do
             {
                 // use signed type to avoid overflowing
-                ::std::int32_t __shifted_local_id = __local_id - __local_id % __k - 1;
+                std::int32_t __shifted_local_id = __local_id - __local_id % __k - 1;
                 if (__shifted_local_id >= 0 && __adjusted_global_id < __n && __local_id % (2 * __k) >= __k &&
                     __local_id % (2 * __k) < 2 * __k - 1)
                 {
@@ -938,39 +958,41 @@ struct __scan
             __wg_assigner(__wg_sums_ptr, __group_id, __local_acc, __local_id);
     }
 
-    template <typename _NDItemId, typename _Size, typename _AccLocal, typename _InAcc, typename _OutAcc,
-              typename _WGSumsPtr, typename _SizePerWG, typename _WGSize, typename _ItersPerWG>
+    template <typename _Size, typename _AccLocal, typename _InAcc, typename _OutAcc, typename _WGSumsPtr>
     void
-    scan_impl(_NDItemId __item, _Size __n, _Size __n_out, _AccLocal& __local_acc, const _InAcc& __acc,
-              _OutAcc& __out_acc, _WGSumsPtr* __wg_sums_ptr, _SizePerWG __size_per_wg, _WGSize __wgroup_size,
-              _ItersPerWG __iters_per_wg, _InitType __init, std::true_type /*has_known_identity*/) const
+    __scan_impl(sycl::nd_item<1> __item, _Size __n, _Size __n_out, _AccLocal& __local_acc, const _InAcc& __acc,
+                _OutAcc& __out_acc, _WGSumsPtr* __wg_sums_ptr, std::size_t __size_per_wg, std::size_t __wgroup_size,
+                std::size_t __iters_per_wg, _InitType __init, std::true_type /*has_known_identity*/) const
     {
-        auto __group_id = __item.get_group(0);
-        auto __local_id = __item.get_local_id(0);
-        auto __use_init = __init_processing<_Tp>{};
+        std::size_t __group_id = __item.get_group(0);
+        std::size_t __local_id = __item.get_local_id(0);
+        __init_processing<_Tp> __use_init{};
 
         constexpr auto __shift = _Inclusive{} ? 0 : 1;
 
         _Size __adjusted_global_id = __local_id + __size_per_wg * __group_id;
-        auto __adder = __local_acc[0];
+        auto __adder = _Tp{__known_identity<_BinaryOperation, _Tp>};
+        if (__group_id == 0)
+            __use_init(__init, __adder, __bin_op);
+
         for (auto __iter = 0; __iter < __iters_per_wg; ++__iter, __adjusted_global_id += __wgroup_size)
         {
             if (__adjusted_global_id < __n)
+            {
                 __local_acc[__local_id] = __data_acc(__adjusted_global_id, __acc);
+                __local_acc[__local_id] = __unary_op(__local_id, __local_acc);
+            }
             else
                 __local_acc[__local_id] = _Tp{__known_identity<_BinaryOperation, _Tp>};
 
-            // the result of __unary_op must be convertible to _Tp
-            _Tp __old_value = __unary_op(__local_id, __local_acc);
-            if (__iter > 0 && __local_id == 0)
-                __old_value = __bin_op(__adder, __old_value);
-            else if (__adjusted_global_id == 0)
-                __use_init(__init, __old_value, __bin_op);
+            // TODO: investigate why the code below sometimes produces incorrect results for copy_if etc.,
+            //       and then decide whether to switch it back on or to keep __log_scan_over_group.
+            // _Tp __value = __local_acc[__local_id];
+            // __local_acc[__local_id] =
+            //    __dpl_sycl::__inclusive_scan_over_group(__item.get_group(), __value, __bin_op, __adder);
+            __local_acc[__local_id] = __log_scan_over_group(__wgroup_size, __item, __local_id, __local_acc, __adder);
 
-            __local_acc[__local_id] =
-                __dpl_sycl::__inclusive_scan_over_group(__item.get_group(), __old_value, __bin_op);
             __dpl_sycl::__group_barrier(__item);
-
             __adder = __local_acc[__wgroup_size - 1];
 
             __gl_assigner(__acc, __out_acc, __adjusted_global_id + __shift, __n, __n_out, __local_acc, __local_id);
@@ -983,15 +1005,14 @@ struct __scan
             __wg_assigner(__wg_sums_ptr, __group_id, __local_acc, __local_id);
     }
 
-    template <typename _NDItemId, typename _Size, typename _AccLocal, typename _InAcc, typename _OutAcc,
-              typename _WGSumsPtr, typename _SizePerWG, typename _WGSize, typename _ItersPerWG>
-    void operator()(_NDItemId __item, _Size __n, _Size __n_out, _AccLocal& __local_acc, const _InAcc& __acc,
-                    _OutAcc& __out_acc, _WGSumsPtr* __wg_sums_ptr, _SizePerWG __size_per_wg, _WGSize __wgroup_size,
-                    _ItersPerWG __iters_per_wg,
-                    _InitType __init = __no_init_value<typename _InitType::__value_type>{}) const
+    template <typename _Size, typename _AccLocal, typename _InAcc, typename _OutAcc, typename _WGSumsPtr>
+    void
+    operator()(sycl::nd_item<1> __item, _Size __n, _Size __n_out, _AccLocal& __local_acc, const _InAcc& __acc,
+               _OutAcc& __out_acc, _WGSumsPtr* __wg_sums_ptr, std::size_t __size_per_wg, std::size_t __wgroup_size,
+               std::size_t __iters_per_wg, _InitType __init = __no_init_value<typename _InitType::__value_type>{}) const
     {
-        scan_impl(__item, __n, __n_out, __local_acc, __acc, __out_acc, __wg_sums_ptr, __size_per_wg, __wgroup_size,
-                  __iters_per_wg, __init, __has_known_identity<_BinaryOperation, _Tp>{});
+        __scan_impl(__item, __n, __n_out, __local_acc, __acc, __out_acc, __wg_sums_ptr, __size_per_wg, __wgroup_size,
+                    __iters_per_wg, __init, __has_known_identity<_BinaryOperation, _Tp>{});
     }
 };