+++ test/parallel_api/ranges/range.zip/zip_view.range.concept.compile.pass.cpp
@@ -321,6 +321,8 @@ void testConceptTuple() {
 }
 
 using OutputIter = cpp17_output_iterator<int*>;
+
+#if !_ONEDPL_CPP20_IN_OUT_ITERATOR_BROKEN
 static_assert(std::output_iterator<OutputIter, int>);
 
 struct OutputView : std::ranges::view_base {
@@ -339,6 +341,7 @@ concept zippable = requires {
 static_assert(!zippable<OutputView>);
 static_assert(!zippable<SimpleCommon, OutputView>);
 static_assert(zippable<SimpleCommon>);
+#endif // !_ONEDPL_CPP20_IN_OUT_ITERATOR_BROKEN
 #endif //_ENABLE_STD_RANGES_TESTING
 
 int main() {