+++ test/parallel_api/ranges/std_ranges_test.h
@@ -103,11 +103,20 @@ struct P2
 {
     P2() {}
     P2(int v): x(v) {}
+    P2(int v, int w): x(v), y(w) {}
     int x = {};
     int y = {};
 
     int proj() const { return x; }
     friend bool operator==(const P2& a, const P2& b) { return a.x == b.x && a.y == b.y; }
+};
+
+struct P3 : public P2
+{
+    using P2::P2;
+    P3(int v): P2(v, v + 13) {}
+
+    friend bool operator==(const P3& a, const P3& b) { return a.x == b.x && a.y == b.y; }
 };
 
 struct A
@@ -480,6 +489,8 @@ public:
 
         //test cases with empty sequence(s)
         process_data_in_in(max_n, 0, 0, CLONE_TEST_POLICY(exec), algo, checker, tr_in, args...);
+        process_data_in_in(max_n, r_size, 0, CLONE_TEST_POLICY(exec), algo, checker, tr_in, args...);
+        process_data_in_in(max_n, 0, r_size, CLONE_TEST_POLICY(exec), algo, checker, tr_in, args...);
     }
 
 private:
@@ -660,8 +671,13 @@ public:
         process_data_in_in_out(max_n, r_size, r_size/2, r_size, CLONE_TEST_POLICY(exec), algo, checker, args...);
         process_data_in_in_out(max_n, r_size, r_size, r_size/2, CLONE_TEST_POLICY(exec), algo, checker, args...);
 
-	    //test cases with empty sequence(s)
+        //test cases with empty sequence(s) and/or zero output capacity
         process_data_in_in_out(max_n, 0, 0, 0, CLONE_TEST_POLICY(exec), algo, checker, args...);
+        process_data_in_in_out(max_n, r_size, r_size, 0, CLONE_TEST_POLICY(exec), algo, checker, args...);
+        process_data_in_in_out(max_n, 0, r_size / 2, r_size, CLONE_TEST_POLICY(exec), algo, checker, args...);
+        process_data_in_in_out(max_n, r_size / 2, 0, r_size, CLONE_TEST_POLICY(exec), algo, checker, args...);
+        process_data_in_in_out(max_n, 0, r_size / 2, r_size / 4, CLONE_TEST_POLICY(exec), algo, checker, args...);
+        process_data_in_in_out(max_n, r_size / 2, 0, r_size / 4, CLONE_TEST_POLICY(exec), algo, checker, args...);
     }
 private: