+++ documentation/release_notes.rst
@@ -15,9 +15,18 @@ Fixed Issues
 ------------
 - Fixed validation of minimal requirements for range-based algorithms. They require clang 16 and newer
   instead of the corresponding libc++ versions.
+- Fixed ``ranges::unique_copy`` to allow output ranges of any size.
+- Fixed the default template argument for the new value type in `ranges::replace` and `ranges::replace_if`
+  to not use projections.
+
+Known Issues and Limitations
+----------------------------
+New in This Release
+^^^^^^^^^^^^^^^^^^^
+- ``ranges::unique_copy`` with the output size smaller than the input size may lose performance on GPU devices.
 
 New in 2022.11.0
-=======================
+================
 
 New Features
 ------------
@@ -39,7 +48,7 @@ Known Issues and Limitations
 ----------------------------
 New in This Release
 ^^^^^^^^^^^^^^^^^^^
-- ``ranges::copy_if`` with the output size smaller than the input size may lose performance on Intel's GPU devices.
+- ``ranges::copy_if`` with the output size smaller than the input size may lose performance on GPU devices.
 
 Existing Issues
 ^^^^^^^^^^^^^^^